web/requirements.txt
fastapi>=0.105.0,<0.106.0
pydantic>=2.5.0,<2.6.0
uvicorn>=0.24.0,<0.25.0
google-auth>=2.0.0,<3.0.0
requests>=2.0.0,<3.0.0
websockets>=12.0,<13.0